backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Palash A.",email = "palash.agarwal.96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2,<4.0.0",
    "flask-socketio>=5.3.6",
    "flask-cors>=4.0.0",
    "chromadb>=0.4.22",
    "pypdf>=3.17.0",
    "sentence-transformers>=2.2.2",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.2",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.25",
    "bcrypt>=4.1.2",
    "python-engineio>=4.9.0",
    "python-socketio>=5.11.0"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
